Support ip -> index
We construct a second radix tree, indexed by network
(e.g. if RADIX_LENGTH is 20, then we make 2**20 entries,
one for each /20). If the IP address is in the radix, this
is an easy lookup. Otherwise, we have to walk the tree to
get this IP's index. Currently, the ip->idx radix adds ~6MB of
memory.